Basic GUI and drawing functionality
Made GUIFrame, which is a basic window which exists to hold the
GUICanvas, the rendering canvas.
Added a Drawable interface with a draw(Graphics, GUICanvas) and
implemented it in Room, Player, Item.
Created Room and Item for testing purposes.
Created a potential file template for loading and creating specific Room
objects.