SW Hallway and fixed desk view scaling
Added placeholder images and parse file for SW Hallway. Fixed desk
scaling by adding switch statements to setScaledImage() method to deal
with the viewDirection and orientation of the desk.
Also added a RoomIndex file to the resources - this would be read into
whatever method was setting up the game and would tell it what rooms to
parse