Tooltips and items now usable!
if you mouseover an item while close enough to it it will display a
tooltip with a description. If you right click on the item while the
tooltip is displayed it will 'use' the item - this does nothing at the
moment aside from the keycard which just prints a line. This works from
any direction