New item functionality and merge
hopefully this merges without issues. The torch can now be turned on and
off by pressing the number of the corresponding inventory slot, the icon
of the item also changes to reflect this. The player.inventoryContains()
method now returns the item if found (in the case of the torch so it can
see if it's on). Fixed mini health bar drawing from different view
directions. Also allowed for scaling of item images if they're inside
other containers