Respawning treasures
Consuming a treasure for points now causes a new treasure to be spawned
on a random itemspawner (not a pouch in case a player is holding it).
Added spawnItem method into the controller to handle this.
Also restructured the initial item spawner method so it doesnt remove
spawners from the collection - so they can be used again
Fixed cabinet and vanishing cabinet view direction bug.