Made TestFrame
TestFrame is a GUIFrame that doesn't open/display itself. It exists for
testing purposes: so when you run a test that involves a GUI in some
way, you can use this one so you don't have 30 windows open by the time
the tests are finished.
TestController uses a TestFrame as its gui.